+++ cpsycle/host/src/keyboardmiscconfig.h
@@ -58,7 +58,7 @@ bool keyboardmiscconfig_advancelineonrecordtweak(const KeyboardMiscConfig*);
 uintptr_t keyboardmiscconfig_pgupdowntype(const KeyboardMiscConfig*);
 intptr_t keyboardmiscconfig_pgupdownstep(const KeyboardMiscConfig*);
 void keyboardmiscconfig_setcursorstep(KeyboardMiscConfig*, int step);
-int keyboardmiscconfig_cursor_step(const KeyboardMiscConfig*);
+intptr_t keyboardmiscconfig_cursor_step(const KeyboardMiscConfig*);
 bool keyboardmiscconfig_ft2fileexplorer(const KeyboardMiscConfig*);
 bool keyboardmiscconfig_following_song(const KeyboardMiscConfig*);
 void keyboardmiscconfig_follow_song(KeyboardMiscConfig*);
